Bug 1267683 - Forbid requests made using default scripting user agents
In the past we've had infra incidents caused by excessive use of
Treeherder's API. To make the accidental cases of these easier to
diagnose, consumers of our API should set an appropriate user agent.

To enforce this, common default user agents are now blacklisted using
the Django common middleware:
https://docs.djangoproject.com/en/1.8/ref/middleware/#module-django.middleware.common
https://docs.djangoproject.com/en/1.8/ref/settings/#disallowed-user-agents

Docs for this change have landed in advance, in #1534.